backend/requirements.txt
# === Core Framework ===
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# === Database ===
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
greenlet==3.1.1

# === Authentication ===
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# === Validation & Settings ===
pydantic==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0

# === Redis & Caching ===
redis[hiredis]==5.2.1

# === Rate Limiting ===
slowapi==0.1.9

# === File Handling ===
python-magic==0.4.27
Pillow==11.0.0

# === HTTP Client ===
httpx==0.28.1

# === Testing ===
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1

# === Dev Tools ===
black==24.10.0
ruff==0.8.4
mypy==1.13.0
